requirements.txt
tkinter
pillow
python-dotenv
sqlite3